pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
authors = [
  {email = "thomas@gibfest.dk"},
  {name = "Thomas Steen Rasmussen"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "exifread==3.0.0",
    "httpx==0.27.2",
    "opencv-python==4.10.0.84",
    "pillow==11.0.0",
    "python-magic==0.4.27",
]
name = "bma-client-lib"
description = "BornHack Media Archive Client Library"
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pre-commit==4.0.0",
    "setuptools-scm==8.0.4",
]
test = [
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
]

[project.urls]
homepage = "https://github.com/bornhack/bma-client-lib-python"

[tool.setuptools_scm]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
extend-exclude = [
    ".git",
    "__pycache__",
]
lint.select = ["ALL"]
lint.ignore = [
    "G004", # https://docs.astral.sh/ruff/rules/logging-f-string/
    "ANN101", # https://docs.astral.sh/ruff/rules/missing-type-self/
    "ANN102", # https://docs.astral.sh/ruff/rules/missing-type-cls/
    "EM101", # https://docs.astral.sh/ruff/rules/raw-string-in-exception/
    "EM102", # https://docs.astral.sh/ruff/rules/f-string-in-exception/
    "COM812", # missing-trailing-comma (https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "ISC001", # single-line-implicit-string-concatenation (https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules)
    "ARG001", # https://docs.astral.sh/ruff/rules/unused-function-argument/
    "ARG002", # https://docs.astral.sh/ruff/rules/unused-method-argument/
    "ARG004", # https://docs.astral.sh/ruff/rules/unused-static-method-argument/
]
line-length = 120

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = [
    "S101", # https://docs.astral.sh/ruff/rules/assert/
    "ANN002", # https://docs.astral.sh/ruff/rules/missing-type-args/
    "ANN003", # https://docs.astral.sh/ruff/rules/missing-type-kwargs/
]

[tool.mypy]
mypy_path = "src"
strict = true
